fix: Translate service session titles
Adds getTranslated(str, locale) to `utils/timetable.ts` that takes str
and tries to translate to locale using built in map. If no translation
is found it just returns str.

getTranslated is applied to all service session titles only as agreed
on Slack. Unfortunately we cannot add anything to JSON as it gets
downloaded from Sessionize API so this is the best we can do.

Service sessions are not clickable so only place where translations
are needed is really TimetableList.